Feat: Expo Router 파일 구조 수정
app 폴더 아래에 있어야 라이브러리 작동
