+++ CookieeApp/app/[...unmatched].js
@@ -1,0 +1,3 @@
+import { Unmatched } from "expo-router";
+
+export default Unmatched;
