--- CookieeApp/app/about.js
@@ -1,10 +1,0 @@
-import { Pressable, Text } from "react-native";
-import { Link } from "expo-router";
-
-export default function about() {
-  return (
-    <Pressable>
-      <Text>about</Text>
-    </Pressable>
-  );
-}
