+++ CookieeApp/app/modal/EventImagePicker.js
