+++ CookieeApp/babel.config.js
@@ -1,7 +1,11 @@
 module.exports = function (api) {
   api.cache(true);
+
   return {
     presets: ["babel-preset-expo"],
-    plugins: ["expo-router/babel"],
+    plugins: [
+      require.resolve("expo-router/babel"),
+      /* */
+    ],
   };
 };
