+++ CookieeApp/package.json
@@ -25,10 +25,10 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-google-button": "^0.7.2",
-    "react-native": "0.72.3",
-    "react-native-gesture-handler": "^2.13.4",
+    "react-native": "0.72.6",
+    "react-native-gesture-handler": "~2.12.0",
     "react-native-google-signin": "^2.1.1",
-    "react-native-reanimated": "^3.5.4",
+    "react-native-reanimated": "~3.3.0",
     "react-native-safe-area-context": "4.6.3",
     "react-native-screens": "~3.22.0",
     "react-native-web": "~0.19.6"
